+++ udata-front-plugins-helper/src/index.js
@@ -1,5 +1,0 @@
-import Pagination from "./pagination.vue";
-import getVisiblePages, { PAGES_AROUND } from "./paginate";
-export * from "./hook";
-
-export { Pagination, getVisiblePages, PAGES_AROUND, };