+++ udata_front/theme/gouvfr/datagouv-components/src/components/QualityComponentInline/index.ts
@@ -1,0 +1,2 @@
+import QualityComponentInline from "./QualityComponentInline.vue";
+export { QualityComponentInline };