+++ udata_front/theme/gouvfr/datagouv-components/src/components/QualityItemWarning/index.ts
@@ -1,0 +1,2 @@
+import QualityItemWarning from "./QualityItemWarning.vue";
+export { QualityItemWarning }