+++ udata_front/theme/gouvfr/datagouv-components/src/components/ReadMore/ReadMore.stories.ts
@@ -1,0 +1,48 @@
+import type { Meta, StoryObj } from '@storybook/vue3';
+import { ReadMore } from ".";
+
+const meta = {
+  title: 'Components/ReadMore',
+  component: ReadMore,
+  tags: ['autodocs'],
+  parameters: {
+    docs: {
+      description: {
+        component: "The ReadMore Component allows to hide the content of a long text and to display it by click on the button."
+      }
+    }
+  },
+} satisfies Meta<typeof ReadMore>;
+
+export default meta;
+
+export const DefaultReadMore: StoryObj<typeof meta> = {
+  render: (args) => ({
+    components: { ReadMore },
+    setup() {
+      return { args };
+    },
+    template: ` <ReadMore v-bind="args" max-height="">
+      <div class="white-space-pre-wrap overflow-wrap-anywhere">
+        Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?
+        Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?
+      </div>
+    </ReadMore>`,
+  }),
+  args: {},
+};
+
+export const NoReadMoreOnSmallContent: StoryObj<typeof meta> = {
+  render: (args) => ({
+    components: { ReadMore },
+    setup() {
+      return { args };
+    },
+    template: ` <ReadMore v-bind="args">
+      <div class="white-space-pre-wrap overflow-wrap-anywhere">
+        Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?
+      </div>
+    </ReadMore>`,
+  }),
+  args: {},
+};