+++ udata_front/theme/gouvfr/datagouv-components/src/components/Well/index.ts
@@ -1,0 +1,4 @@
+import "../well.less";
+import Well from "./Well.vue";
+
+export default Well;