+++ udata_front/theme/gouvfr/datagouv-components/src/components/index.ts
@@ -1,5 +1,14 @@
 import { CopyButton } from "./CopyButton";
 import OrganizationNameWithCertificate from "./Organization";
+import Pagination from "./Pagination";
 import { ResourceAccordion, Loader as ResourceAccordionLoader } from "./ResourceAccordion";
+import { QualityItem } from "./QualityItem";
+import { QualityItemWarning } from "./QualityItemWarning";
+import { QualityScore } from "./QualityScore";
+import { Toggletip } from "./Toggletip";
+import { QualityComponent } from "./QualityComponent"
+import { QualityComponentInline } from "./QualityComponentInline"
+import { ReadMore } from "./ReadMore"
+import Well from "./Well";
 
-export { CopyButton, OrganizationNameWithCertificate, ResourceAccordion, ResourceAccordionLoader };
+export { CopyButton, OrganizationNameWithCertificate, Pagination, QualityComponent, QualityComponentInline, QualityItem, QualityItemWarning, QualityScore, ReadMore, ResourceAccordion, ResourceAccordionLoader, Toggletip, Well };